feat(main.py,-ChatWindow.tsx): establish new connection between client and server to do audio transcription and text to speech
Implemented 2 new endpoints in the FastAPI backend for doing transcription and TTS. Modified the transcribeAndSendMessage and playMessageAudio functions to call the backend for doing those tasks. Also, added marginLeft for the <AutoResizeTextarea> component to avoid overlap with the microphone icon in the UI. Actual implementation of hosting a custom TTS model is for later.

BREAKING CHANGE:


Former-commit-id: e33b3d7711314783c64ece088ce3ef2d888cfc1b